scale monster in calc
This is a fairly involved change to the way we scale monsters in a few ways to resolve #210. The underlying issue was that scaling was being called twice, once to determine the base monster and once to apply the hp change.

The changes made were:
1. Break up scaleMonster into its components so that Vardorvis HP scaling can be handled independently, and also to make each scaling facet more modularized.
2. Move scaleMonster calls into the calculator constructor, with a specific flag to disable this option. The flag is needed to allow the loadout comparison to alter the stats *after* scaling, without triggering a re-scale.
3. In calc's distAtHp, use scaleMonsterHpOnly and skip scaling in the sub-calculator's constructor.
4. Added a new noInit internal flag to the calculator which allows a parent calculator to skip monster scaling, equipment checks, etc. that it has already performed.

Closes #210
